fix: add proper help text for mirror command using -h flag
